back/requirements.txt
appnope==0.1.0
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
decorator==4.4.2
Django==1.11.29
django-cors-headers==3.0.2
djangorestframework==3.9.4
enum34==1.1.10
ipdb==0.13.4
ipython==5.10.0
ipython-genutils==0.2.0
pathlib2==2.3.5
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==1.0.18
ptyprocess==0.6.0
Pygments==2.5.2
pytz==2020.1
scandir==1.10.0
simplegeneric==0.8.1
six==1.15.0
traitlets==4.3.3
wcwidth==0.2.5
